+++ doc/changes.rst
@@ -5,6 +5,7 @@ desitarget Change Log
 1.1.0 (unreleased)
 ------------------
 
+* Support reading mtl 1.0.0 format with different columns [`PR #742`_].
 * Enable optional subpriority overrides [`PR #740`_, `PR #741`_].
 * Allow initial ledgers to use a preordained timestamp [`PR #739`_].
     * ``MWS_FAINT`` targets can be exempted from this timestamp.
@@ -36,6 +37,7 @@ desitarget Change Log
 .. _`PR #739`: https://github.com/desihub/desitarget/pull/739
 .. _`PR #740`: https://github.com/desihub/desitarget/pull/740
 .. _`PR #741`: https://github.com/desihub/desitarget/pull/741
+.. _`PR #742`: https://github.com/desihub/desitarget/pull/742
 
 1.0.1 (2021-05-14)
 ------------------