+++ doc/changes.rst
@@ -5,6 +5,8 @@ desitarget Change Log
 1.1.0 (unreleased)
 ------------------
 
+* override_subpriority support duplicate input TARGETID from secondaries
+  [`PR #743`_].
 * Support reading mtl 1.0.0 format with different columns [`PR #742`_].
 * Enable optional subpriority overrides [`PR #740`_, `PR #741`_].
 * Allow initial ledgers to use a preordained timestamp [`PR #739`_].
@@ -38,6 +40,7 @@ desitarget Change Log
 .. _`PR #740`: https://github.com/desihub/desitarget/pull/740
 .. _`PR #741`: https://github.com/desihub/desitarget/pull/741
 .. _`PR #742`: https://github.com/desihub/desitarget/pull/742
+.. _`PR #743`: https://github.com/desihub/desitarget/pull/743
 
 1.0.1 (2021-05-14)
 ------------------